overlays: Move vertex & vector utility classes to new file
